HHH-14642 revert back BasicCriteriaUsageTest so its testDateCompositeCustomType() would not be lost (it becomes outdated due to the code changes to the Entity)
